+++ src/components/Home/About/ServerListAbout.tsx
@@ -3,7 +3,7 @@ import aboutServices from "@/services/aboutServices";
 import { aboutProps } from "@/types/about";
 import Image from "next/image";
 import { GITHUB, LINKEDIN, MAIL, MEDIUM } from "@/constant/urls";
-import Photo from "../Photo";
+import Photo from "../../ui/Photo";
 const dimensions = {
   height: 0,
   width: 0,
