+++ src/components/ui/Photo.tsx
